Move themed components to the right folder
