+++ src/components/scenes/FioRequestListScene.js
@@ -24,10 +24,10 @@ import type { FioRequest, GuiWallet } from '../../types/types'
 import { scale } from '../../util/scaling.js'
 import FullScreenLoader from '../common/FullScreenLoader'
 import { SceneWrapper } from '../common/SceneWrapper'
-import { SettingsHeaderRow } from '../common/SettingsHeaderRow.js'
 import type { WalletListResult } from '../modals/WalletListModal'
 import { WalletListModal } from '../modals/WalletListModal'
-import { Airship, showError } from '../services/AirshipInstance'
+import { Airship, showError } from '../services/AirshipInstance.js'
+import { SettingsHeaderRow } from '../themed/SettingsHeaderRow.js'
 
 const SCROLL_THRESHOLD = 0.5
